Move model components and loss functions to models and training, respectively
