+++ src/graphnet/models/components/pool.py
