+++ src/graphnet/models/gnn/dynedge.py
@@ -11,7 +11,7 @@ from torch import Tensor
 from torch_geometric.data import Data
 from torch_geometric.nn import EdgeConv
 from torch_scatter import scatter_max, scatter_mean, scatter_min, scatter_sum
-from graphnet.components.layers import DynEdgeConv
+from graphnet.models.components.layers import DynEdgeConv
 from graphnet.models.coarsening import Coarsening
 
 from graphnet.models.gnn.gnn import GNN
