+++ src/graphnet/models/task/task.py
@@ -17,7 +17,7 @@ from torch import Tensor
 from torch.nn import Linear
 from torch_geometric.data import Data
 
-from graphnet.components.loss_functions import LossFunction
+from graphnet.training.loss_functions import LossFunction
 
 
 class Task(LightningModule):
