+++ tests/ml/test_loss_functions.py
@@ -6,7 +6,7 @@ import torch
 from torch import Tensor
 from torch.autograd import grad
 
-from graphnet.components.loss_functions import LogCoshLoss, VonMisesFisherLoss
+from graphnet.training.loss_functions import LogCoshLoss, VonMisesFisherLoss
 from graphnet.utilities.maths import eps_like
 from torch_geometric.data import Data
